middleware/requirements.txt
Flask==3.0.3
requests==2.32.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
